+++ packages/health/lib/src/data_types.dart
@@ -24,10 +24,12 @@ enum HealthDataType {
   MOVE_MINUTES,
   DISTANCE_DELTA,
   MINDFULNESS,
-  SLEEP_ANALYSIS,
   WATER,
+  SLEEP_IN_BED,
+  SLEEP_ASLEEP,
+  SLEEP_AWAKE,
 
-  // Heart Rate events (specific to Apple Watch)
+// Heart Rate events (specific to Apple Watch)
   HIGH_HEART_RATE_EVENT,
   LOW_HEART_RATE_EVENT,
   IRREGULAR_HEART_RATE_EVENT,
@@ -60,7 +62,9 @@ const List<HealthDataType> _dataTypeKeysIOS = [
   HealthDataType.FLIGHTS_CLIMBED,
   HealthDataType.DISTANCE_WALKING_RUNNING,
   HealthDataType.MINDFULNESS,
-  HealthDataType.SLEEP_ANALYSIS,
+  HealthDataType.SLEEP_IN_BED,
+  HealthDataType.SLEEP_AWAKE,
+  HealthDataType.SLEEP_ASLEEP,
   HealthDataType.WATER,
 ];
 
@@ -106,6 +110,14 @@ const Map<HealthDataType, HealthDataUnit> _dataTypeToUnit = {
   HealthDataType.FLIGHTS_CLIMBED: HealthDataUnit.COUNT,
   HealthDataType.MOVE_MINUTES: HealthDataUnit.MINUTES,
   HealthDataType.DISTANCE_DELTA: HealthDataUnit.METERS,
+
+  HealthDataType.WATER: HealthDataUnit.LITER,
+  HealthDataType.SLEEP_IN_BED: HealthDataUnit.MINUTES,
+  HealthDataType.SLEEP_ASLEEP: HealthDataUnit.MINUTES,
+  HealthDataType.SLEEP_AWAKE: HealthDataUnit.MINUTES,
+  HealthDataType.MINDFULNESS: HealthDataUnit.MINUTES,
+
+  /// Heart Rate events (specific to Apple Watch)
   HealthDataType.HIGH_HEART_RATE_EVENT: HealthDataUnit.NO_UNIT,
   HealthDataType.LOW_HEART_RATE_EVENT: HealthDataUnit.NO_UNIT,
   HealthDataType.IRREGULAR_HEART_RATE_EVENT: HealthDataUnit.NO_UNIT,