+++ packages/weather/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 1.2.3
+* Now uses HTTPS rather than HTTP for making API calls
+
 ## 1.2.2
 * Added wind gust